requirements.txt
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
dropbox==12.0.2
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.3.2
oauthlib==3.3.1
pandas==2.3.1
ply==3.11
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
six==1.17.0
stone==3.3.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
